requirements.txt
gradio
openai
pillow
python-dotenv
requests
